requirements.txt
numpy>=1.21.0
matplotlib>=3.4.0
requests>=2.26.0
python-dotenv>=0.19.0
torch>=2.0.0
pytest>=7.0.0
tensorboard>=2.15.0
tqdm>=4.65.0 